Remove unused code, move Model.Base -> GH.Model
The types in Model.Base were all GitHub specific, so put them in
GitHub.Model. Even if we start using these types in a model definition,
we can import GitHub.Model there because the GitHub modules should not
ever depend on application code (so we'll avoid the common cyclic
imports through Model).