Remove extracted modules, continuous deployment
As part of this we moved to the actual github Hackage library, with
minimally duplicated extensions. We could work to remove even these via
upstreaming or relying more on primitives at the CLI boundary. Currently
we want to use typed App and Installation Ids, which requires our Data
extensions in both places.