Add authorization, admins, and admin/signups
Now that our database is correctly fire-walled, I need an in-app way to
interact with the data we store, which is currently just signups.

A simple admin/signups page will do, but still required some way to
identify myself as an admin, which required proper authentication with
GitHub, which required updated yesod-auth-oauth2, and turned out to
still need a custom GitHubApps-specific plugin.