Refactor Backend, store Job data
Backend should be an isolated namespace for actions that don't require
Handler (they only need IO and Logger). This namespace is similarly-
structured to the web-app, and the same AppSettings type is used to
reduce duplication and aid discover-ability.

Access to Redis and the DB is available in the Backend, but interfaces
with Handler should be minimized and used only when an actual Handler
needs to pass data to the backend (e.g. enqueueRestylerJob).

In the future, we may generalize the Job pattern to support distinct
jobs on distinct queues (ideally, abstracting over Redis as an
implementation detail) with queue-specific Backend "Handlers".