Fix deployment tag
The image and deploy job are different build numbers, so we either need
to find another variable that's consistent across them, or persist the
value from the build/release to the deploy job.

This implements the latter. It seems overly-complicated, but there are
no good candidates for the former, so we'll do it for now.