Improve OwnPullRequest webhook predicate
In addition to looking for -restyled as a branch suffix, confirm the
author also looks like us. We may only need the author check, and in
fact, we may change our branch scheme some day and break this, but we'll
keep both for now to be conservative in our changes.

Fixes #86.