Ensure proper persistent logging in the Backend
Persistent does something kind of clever: the functions that create the
connections to begin with (e.g. createPostgresqlPool) require
MonadLogger. But the functions that actually run queries (e.g.
runSqlPool) don't -- and yet it's the latter that actually logs. It
turns out the logging context you provide to the creation function is
captured and used by the execution functions. In other words: it matters
for more than just creation-time log messages.

I did not know this, and was just using runStdoutLoggingT to create the
pool because, who cares what's logged then? But this meant that that's
how all persistent logging happened in the Backend. Specifically, there
was not level-based filtering. Extracting and re-using our level-aware
logging functions fixes that.

Additionally, I moved to the more conventional SqlPersistT interface.