Let users exist without an email
Despite requesting the user:email scope, we are not always receiving
emails back in our authorization responses from GitHub.

It's hard to say definitively what causes this. One theory is that
scope-handling in GitHub App OAuth apps has a bug and it only returns
public emails even though the scope authorizes you to get the private
one too. Another is that because the affected users had recently changed
domains at their company, but still had their GitHub profile linked to
the old domain, that is the root cause.

Either way: it's possible, so let's handle it.

I'm also not sure if we can always expect emails from GitLab or not, but
since we have to handle it for GitHub too, might as well keep things
consistent.