Copy config/revision directly from .git
This assumes we only care about this value when we're building
production images from master. So we can just copy that ref's sha
content. To be more accurate, we would need to either manually follow
the indirection of refs/HEAD or COPY in all of .git and use an actual
git-rev-parse. Neither seems worth it so far.