Refactor deployment approach
We can build a single image on Quay, then construct dedicated
Dockerfiles (with specific CMD) in the deploy job itself. This will be
twice as fast and was in fact how we were doing it before. It also means
we release both processes at once.

I was originally trying to remove the dependency on the heroku toolbelt,
as well as not build images on Circle, but:

- We need heroku anyway to release
- As long as it's just FROM/CMD, it should be fast enough even un-cached
  (definitely faster than waiting on two images form Quay)