Change cached stack paths to match container
The .stack-work director is not relocatable, and the checkout path
includes the agent id. This should function, but means caches are
agent-local. Perhaps that'll be a nice middle ground anyway.

And if we're going to match that one, we might as well match the .stack
mount as well. That /cache directory was arbitrary anyway.