Stop using .env
docker-compose shipped some kind of update where now it's unable to
parse a .env file with multi-line values. We weren't using .env for
docker-compose, but now that it actively errors on this file, with no
way to disable it, we're forced to do something else. Typical.

As part of this we centralized the loadEnv calls in Settings and ensured
we don't load the development .env file (now named .env.development)
unless we're CPP-guarded for DEVELOPMENT. We had a lot of things in
place to avoid shipping and using development settings in production,
but this adds yet another.

Since DEVELOPMENT is technically active in test, we use a dedicated load
function for that, to avoid surprises and keep it simple.