Move Rerun logic and reduce admin routes
Generally, let Admin link to normal pages where ever possible. This
avoids the extra complexity (and security risk) of a large admin area of
the site. The only admin action we had was rerurn, which we want users
to have access to anyway (Closes #135). Any other actions will (for now)
be accomplished via a direct database connection, when required.

This will require authorizing admins on private user-facing pages, which
will follow next.