ErrorFocuser
If we're showing a single card, and there is a log-line with "We had
trouble...", scroll the window and pre element to focus directly only
that content.

Sometimes the real error can get lost with the stream prefixes and
general Job logging that happens before it. Removing the streams and
adding louder decoration of the beginning of the error could be another,
simpler, option, but:

- The streams are valuable IMO, and
- Louder decoration would require coordination with Restyler

So let's try this for now.