Relax bot filter to only ignore Restyled PRs
This was originally implemented to prevent infinite Restyled PRs if the
Restyling somehow always produces a diff (it happens, trust me). It
eventually became ignoring *any* bot user, since that seemed reasonable
and made some users happy.

However, this interacts poorly with setting Restyled as a required
status. Since this mechanism ignores the PR entirely, they don't get a
Green status from Restyled when ignored.

We may change that: defer this check to Restyler and let it report a
Green status for ignored PRs, but that would be higher effort and has
been deferred. For now, we change this back to specifically ignore
Restyled PRs, which are by definition non-master-based and so wouldn't
(normally) have required statuses enabled.

Closes restyled-io/restyler#63.