Test and fix bug in withExtraEnvVars
This resulted in our RestyleMachine selection not working. We would find
the machine, but this bit failed to override DOCKER_ variables that were
pre-existing, set in entrypoint from the days before we had this system.

This meant restyled-prod-2 was forever used, even though we've been
attempting to rotate machines and automate out-of-space issues.

:confused: :facepalm: