Vastly simplify RIO.Process.Follow
Implement the following function in a much more idiomatic style, which
is to accept a ProcessConfig argument, instead of actually calling proc
internally and therefore having to live in an RIO env ourselves.

This composes much better and allows followProcess to be used
interchangeably with other analogous functions, such as readProcess.

Similarly, instead of owning running something on a RestyleMachine, and
therefore needing to call proc, and therefore needing that un-idiomatic
interface, simplify that function to just ENV-modification, which is all
it should be. This lead us to moving our (tested) helper and perhaps a
muddier-than-necessary diff, but oh well.

Together, all of this means our "capture" versions are unnecessary and
we can just use readProcess when that's what we need or followProcess
when that's what we want. :chefs_kiss: