Simplify HealthCheck, avoid multiple alerts
Making rate a Maybe means we can avoid alerting on both 0 completions
and rate when no Jobs have occurred in the hour. However, this presented
a problem with the Show instance, since "stat=Just x" wouldn't parse
correctly in logDNA. To resolve this and keep the stat type variable
totally generic would be far too complex. Instead, we make stat a closed
sum type of fully concrete values that we can operate on and string-ify
as needed. Removing the type variable from HealthCheck also meant we can
put them in a list to traverse_ over, a minor but nice win.