Run restyler containers --detached
Currently, deployments may cause restyle jobs to die or hang. This is a
step in resolving that.

With this change, the Restyler job itself will not die. It will always
finish its work and provide its user value (a Restyled PR). The only
thing "lost" is the Job tracking (i.e. exitCode and completedAt). This
is more for administration than anything, so it's OK to follow up and
reconcile those Jobs later, which this also includes code to support.