Rewrite Job Reconciliation approach
The previous approach basically started with open Jobs then looked for a
StoppedContainer to reconcile with. This broke if the process had
crashed after launching the container, but before writing the
machine/container back to the Job record. It also required the
machine/container columns to begin with.

This approach starts by finding stopped containers, which are labeled
with the job-id when created, then going back to reconcile with an open
Job if we have one. Much more robust and needs no extra data in our
models.