Fix runHealthCheckOnHours
In my head, using a guard here was a convenient syntax for giving
"false" from this check. Not quite right... Because I was in the Maybe
Monad (since some operations in this check are fail-able), the guard
really meant Nothing.

Some other fool (me) decided to default the result of this Maybe check
to True on the Nothing case as an attempt to "fail open" -- i.e. if we
have any trouble knowing if we're in working hours or not, still run the
HealthCheck.

This is a weird mish-mash of both Boolean- and Maybe-blindness, as well
as just mis-use of the short-circuit semantics of Maybe (how do you
"throw" a Just?), and ultimately meant that HealthCheck always ran (at
least "fail open" works).

Purely in an effort to debug, I upgraded the Maybe to Either, which
meant we could attach meaning to "failure" cases. Translating this
naturally removed the guard, changed the throw semantics to match my
desire, and so fixed the bug before I really tracked it down.

From there I collapsed to a flat enumeration of possible results,
removing any significance of Left-Right and making ExceptT an
implementation detail only.

The resulting code should be clearer and harder to misuse. It also
works; though I've elected not add unit tests yet.