Add jobLog column for better log compression
Writing the list of JobLogLine into the deprecated stdout/stderr columns
was fast, but lossy and more complicated to deal with at rendering time.

This approach puts a JSON column of the original JobLogLines themselves.
Compression is a similar process except not lossy, and rendering can
just render them as-is exactly the same way as if they were in a
separate table.

The JSONB data is slightly more, but we'll see how that shakes out. I'm
hitting row limits, not size limits.

This commit implements the new column and support for rendering from it
if/when it's populated. The next commit will actually compress our Jobs
into it, which will be used locally once production is ready to handle
such data.