Ignore PRs from pull[bot]
pull[bot] is a bot to keep Forks up to date. It's unlikely anyone would
want to keep their fork up to date *and* restyle on top of it, that's
just a recipe for merge conflicts later. These PRs are typically also
massive and so tie up Restyle Machines needlessly.

It would be better to ignore this after accepting the Webhook, so we
could set a skipped-job and provide feedback to users. In order to
accomplish this, we'll need to send author information deeping into the
system, which is non-trivial. Therefore, this is being deferred for now.