Ignore Stale Jobs
Before running a Job, check if other concurrent Jobs for the same Pull
Request are in-flight. If such Jobs are newer than ourselves, we should
not run. If such a Job is older than ourselves, we should cancel it.

We don't currently act on the latter. We send the signal, but the
Restyler CLI will ignore it. We are just going to use the log message to
see how often this may happen (if it's worth building that support).