Limit the maximums Jobs on one Restyle Machine
Scaling actions, if not carefully coordinated, can result in a high
services count and a low machines count. If 8 services send Jobs to 1
machine, it gets overloaded and hangs. This is disruptive and requires
manual intervention, sometimes as drastic as restarting the Heroku
service too!

We have some of the coordination in place now, but it's better to
prevent this at the source by limiting the maximum Jobs per machine.

For now, this is a static setting defined for our typical machines size,
but we could let agents submit a maximum for themselves and store it in
restyle_machines in the future.