Catch Plan sync errors, maintain current Accounts
This ensures we synchronize what Plans we can. Apparently GitHub is not
particularly reliable and has been 500'ing on a particular plan for a
while. The plan that's erroring is a free one, so we don't need to
synchronize it, but if the error blocks synchronizing paid Plans, that'd
be an issue.