Remove strict pluralize, provide pluralizeWith
We don't actually have to care about Lazy vs Strict much in this case
because:

- The input is OverloadedStrings (and a numeric)
- The output can be given to display (or #{}) either way

Therefore, we should just use the single definition.

As part of this pluralizeWith was extracted so that we can easily choose
how the number itself is formatted (e.g. with commas or not).

I also updated the imports to treat Formatters as qualified.