Make the range select a little safer
We're using runInputGet to read the range select, which is simple and
convenient, but building the input HTML yourself is annoying because
selectField expects numerically-indexed options, which it handles
internally if you're using the heavier proper Form rendering.

I might just move to that, but this makes it a little less likely to
mess up. At least the introduced error will be noisier.