Extend support for non-GitHub Plans
Previously, we had the one special Plan with github_id=0 to represent
our only non-GitHub Plan. We also had a case statement based on
hard-coded github_id values, to decide the private-repo allowance for
any given Plan.

As of this patch,

1. github_id is NULL-able. No Id means non-GitHub
2. private_repo_allowance is explicit, on the table itself

This means github_id is only used during sync, and not logically within
the application. And we read private_repo_allowance to know how to limit
things for a given Plan.

Now, we can create more non-GitHub plans with their own limitations,
which we'll need to do when we get listed in AppSumo.

NOTE: private_repo_allowance is modeled as a number in the Database,
where negative numbers means Unlimited. This wart is dispatched
immediately through the PersistField instance, so it shouldn't be
problematic, provided we don't grow more applications that directly
interface with the Database.