Fix handling of Friends & Family plan during sync
I *thought* I understood the fix of !force on the Unique index of a
nullable column. I knew that, during the Marketplace sync, we needed to
use that index, but never with a NULL value.

I missed the fact that find-or-create of the F&F Plan was also using
this index, and it was supplying a NULL value to the operation. This
resulted in two negative outcomes:

- It continued to create more and more F&F Plans, every sync

  The UPSERT with NULL was just INSERT.

- It wiped out all the existing F&F Accounts

  Since we got back a new id every time, all /other/ accounts that
  weren't just synchronized got wiped. I hate that this logic exists, it
  was only a matter of time. Once we're fully away from github_id=0
  being special, it'll be harder to have this accidental data loss.

This patches fixes both, though the data can likely not be recovered.