Stop checking that Account matches User for MP claim
Limited plans can be installed for Users or Organizations. If installed
for an Organization, it was previously impossible to change which
repository was enabled. That's because the code as it was asserted that
the logged-in User's Account was present and matched the owner of the
repository. This would only work in User installations.

The code now allows any Use with access to the repository in question to
enable/disable with Restyled. Since this is broadening access, some
logging was added in case we need to audit.

This does allow the following scenario:

1. User 1 is an Admin of Org A
2. User 2 is a non-Admin member of Org A
3. User 1 installs Restyled in Org A
4. User 2 can enable/disable any Repositories they can see from Org A

This feels problematic, but because it's still impossible to for User 2
to incur unexpected costs, we accept it. To do otherwise would require
much complexity on our end.