One-ff script to give back lost F&F accounts
We were able to recover the deleted accounts by locating any private
repositories that had successful Jobs despite not currently having an
Account. These are mostly like the F&F Accounts. There may be a few
folks who previously paid and then churned, but we accept this in order
to recover the (mostly GitHub Student) accounts we almost lost.