Add button to manually reconcile Restyle Machines
We perform reconciliation at two times:

- When a new Webhooks task starts
- When a Restyle Machine de-registers via DELETE /admin/machines/:id

In a Scale Down, neither of these reconciliations will happen for the
single Restyle Machine being left in service.

Therefore, if a scaled-down Webhooks task happens to leave Jobs running
on that Restyle Machine, they won't get reconciled until (most likely)
the next Scale Up event -- which could be a while.

Until then, this state has two negative effects:

- The Build log for such Jobs will appear frozen. Even though the Job
  would complete normally and update PRs accordingly, the log will stay
  static at the point of scale-down, until reconciliation writes it out.

- We operate at incorrect Restyle Machine capacity. The maximum
  job_count and scaling counts is set such that the number of Webhooks
  tasks we run matches the number of total slots available. When we
  mistakenly leave a slot (or more) held, we see "No Restyle Machine
  found" delays in our logs while Webhooks task wait for the
  unreconciled Jobs that we think are still running. One good thing is
  this will lead to a backup and Scale Up and correct the problem,
  eventually.

Long-story-longer, until we can address this, it's nice to have an
explicit and safe button to push to address this situation by
reconciling on-demand.