Prevent spinning in Pager for CloudWatch Events
An empty log seems to spin us forever until throttled.

Interestingly, viewing the stream in the AWS Console has the exact same
behavior. I wonder if this is a bug in responses for expired streams (vs
just always having been empty). I would've expected an empty log to have
no valid nextForwardToken, which would similarly halt pagination.

Anyway, we can easily avoid the spins by ending pagination on an empty
page. This should be fine for the case where logs are streaming in
because we loop correctly on that (for in-progress Jobs) from the
outside.