Refactor and improve log streaming
Always render the page and stream in logs later; we're now js-required,
but we can add a no-js bit of copy soon to point users at the Download
Log link.

Re-write the log-streaming to be fully Conduit from end to end and
cleanup other implementation details (e.g. MonadAWS).

Tweak the UI to remove the "stream" column, lower-case and
better-decorate levels, and mute the colors of non-logging lines
slightly.