Remove legacy Job output fields
We place a 30 day expiry on CloudWatch-based logs. Since we've had that
system in place for new Jobs for close to 30 days, they are going to
expire. If that's acceptable, it's also acceptable to stop reading back
through job.log or job.std(err|out) storage for even older cases.

Closes #278.