Don't attempt to fetch old logs
We know we only retain logs for 30 days. I expected it to be fine to
attempt to fetch them and ignore the exceptions, but we are getting
spikes of request timeouts, often on /logs routes for very old Jobs.

This is an experiment to see if removing the attempts we know won't
succeed is helpful. The downside is we are hard-coding the log-retention
value between our infra and this application.