Move MonadCache to in-memory, not Redis
We're experiencing intermittent 503s on routes that don't have to do a
whole lot (loading 1 Job with an expired log). Pretty much the only work
being done is talking to Redis for caching of authorization results.

We even seen 503s on /system/metrics, which is *definitely* doing
nothing more than talking to Redis.

So Redis is sus and we should experiment with moving away from it.

We don't need robustness or performance of caching across processes, so
moving this to in-memory should be acceptable. This caching really only
had two purposes:

1. To avoid a timing attack by (uncached) authorization of private
   repositories against GitHub APIs from giving away their existence,
   and
2. So we can pre-cache data to testing

Both of these are adequately solved though in-memory caching.

If we can then move from Redis to SQS for the webhooks queue, we can
remove Redis entirely.