Implement Web tracing
This is an attempt to figure out why we're seeing daily timeouts.

Introduces a Middleware to set up a Web Transaction for the Request,
trace a Segment per Middleware, the Handler, both Authz phases, runDB,
runRedis, and GitHub API calls. We attach the http.status to the trace
and use noticeError for 5XX.

Deferred for now:

- A SQL segment with actual query details (a "Datastore" segment)
- Making the GitHub segments "External"

Some interaction between tracing-newrelic's Setup.hs and stack seems to
make it so only a local-directory dependency works correctly w.r.t. the
libnewrelic.so shared library. bin/setup-tracing-newrelic can be used
locally, on CI, and in the Docker image build to get this right. In the
future, we probably just want to require the library is present on the
system and figure out how to ensure that in all the contexts.