Orient App logging around LoggerSet
Yesod is LoggerSet-focused through its own Logger type. It's possible
(and recommend) to stash one on App for re-use. Once there, we can use
it for all of our logging needs to ensure maximum efficiency.

By creating our own Logger type that also holds the LogLevel, we need
only that to perform all logging according to settings provided at
Logger construction. This opens the seam for later, so that we can
adjust other details such as destination, color, and format.