Add jobs/:id/patch
Log messages with context.patch:true are assumed to be the actual
git-format-patch output (soon to be) produced by Restyler. These lines
should be omitted from the normal log view and accessible at this new
route.

This makes it easy for users to do,

    curl .../patch | git am

And get all the Restyling commits applied (complete with author info,
timestamps, etc) locally and pushable directly to their PR.

This workflow is important for forks, where we're unable to create the
sibling PR without risk of bringing untrusted commits of untrusted code
into a non-fork PR (which grants the Actions that run special
privileges).

Since Restyler isn't deployed quite yet, this will add a small "Download
patch" link that goes to an empty page. This should be pretty temporary.