Reset expiresAt when synchronizing GH plans
Offer- and Student-based Plans come with expiry. If this user converts
to a GH plan, it needs to be reset. To be totally correct, we should
only do this if planId is changing, but that's annoying to do and we
know that GH plans never have expiry, so we can just always reset it.
This will also correct any existing cases from before we did this.

There are only such cases for the OSS plan right now, so this is not
affecting any users -- but it would, if an Offer- or Student-based user
converted to a paid plan and we continued to respect this old expiresAt.

Oops.