Support Redis TLS
Heroku is about to force-upgrade us to 6.2, which requires TLS.

It uses a self-signed cert and doesn't authenticate clients (so, like, why even?) so we match
that setup in the local docker-compose. We have to retain non-TLS on CI, since I don't know
how to get the GitHub service to run with TLS. We also configure the local agent for TLS,
even though the agent updates to support it are in progress.