Fix non-JSON log parsing
When implementing support for act logs, we pushed anything that didn't
parse to `trace` level since that stuff seemed _really_ chatty. We
intended to update the UI with a trace toggle (like we have for debug),
but we never go there.

We stopped short of using act in production anyway, so we don't need
that. And the lack of a toggle meant these logs were now just completely
hidden.

While we're in here, we made the prefix parsing a little more robust for
parsing prefixes we see in underlying tools' non-JSON logs too.